% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-jsm2022.Rmd.
\name{get_talks}
\alias{get_talks}
\title{Get talks based on some criteria}
\usage{
get_talks(speakers, authors, people, keywords, days, session_types)
}
\arguments{
\item{speakers}{character vector of speaker names.  Names start with given (first) name, following the format of JSM program.}

\item{authors}{character vector of author names. Same as above.}

\item{people}{character vector of people names.  This looks for the person as a speaker, author, organizer, and chair.}

\item{keywords}{character vector of words to look for in talk title}

\item{days}{vector of Dates (e.g. "2022-08-07")}

\item{session_types}{character vector, e.g. "Invited Papers"}
}
\description{
Get talks based on some criteria
}