% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-jsm2022.Rmd.
\name{ordered_nz}
\alias{ordered_nz}
\title{Get indices of nonzero elements of a vector ordered from largest to smallest}
\usage{
ordered_nz(vec)
}
\arguments{
\item{vec}{a numerical vector}
}
\description{
This is like \code{which(vec != 0)}, but with indices sorted by size of \code{vec}.
}