% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-jsm2022.Rmd.
\docType{data}
\name{words}
\alias{words}
\title{JSM words used in titles}
\format{
A character vector of 3966 JSM author names
}
\source{
This data was scraped from the JSM 2022 program.
}
\usage{
words
}
\description{
JSM words used in titles
}
\keyword{datasets}